refactor(portal-v3): replace legacy action detail in place
