--- web/scenes/PortalV3/Teams/TeamId/Apps/AppId/WorldIdActions/ActionId/Danger/page/index.tsx
@@ -1,85 +1,0 @@
-"use client";
-
-import { SizingWrapper } from "@/components/SizingWrapper";
-import { ActionDangerZone } from "@/components/ActionDangerZone";
-import { ErrorPage } from "@/components/ErrorPage";
-import { urls } from "@/lib/urls";
-import { useRouter } from "next/navigation";
-import { useCallback, useState, use } from "react";
-import { toast } from "react-toastify";
-import { useGetSingleActionV4Query } from "@/scenes/common/Teams/TeamId/Apps/AppId/WorldIdActions/ActionId/page/graphql/client/get-single-action-v4.generated";
-import { deleteActionV4ServerSide } from "./server";
-
-type WorldIdActionIdDangerPageProps = {
-  params: Promise<Record<string, string>>;
-};
-
-export const WorldIdActionIdDangerPage = (
-  props: WorldIdActionIdDangerPageProps,
-) => {
-  const params = use(props.params);
-  const actionId = params?.actionId;
-  const teamId = params?.teamId;
-  const appId = params?.appId;
-  const router = useRouter();
-  const [isDeleting, setIsDeleting] = useState(false);
-
-  const { data, loading, error } = useGetSingleActionV4Query({
-    variables: { action_id: actionId ?? "" },
-  });
-
-  const action = data?.action_v4_by_pk;
-
-  const handleDelete = useCallback(async () => {
-    if (!actionId || !appId || !teamId) {
-      throw new Error("Missing required parameters");
-    }
-
-    setIsDeleting(true);
-    try {
-      const result = await deleteActionV4ServerSide(actionId, appId);
-
-      if (!result.success) {
-        throw new Error(result.message || "Failed to delete action");
-      }
-
-      toast.success("Action deleted successfully");
-      router.push(urls.worldIdActions({ team_id: teamId, app_id: appId }));
-    } catch (error) {
-      throw error;
-    } finally {
-      setIsDeleting(false);
-    }
-  }, [actionId, appId, teamId, router]);
-
-  if (error) {
-    return (
-      <SizingWrapper gridClassName="order-1 md:order-2">
-        <ErrorPage statusCode={500} title="Failed to load action" />
-      </SizingWrapper>
-    );
-  }
-
-  if (!loading && !action) {
-    return (
-      <SizingWrapper gridClassName="order-1 md:order-2">
-        <ErrorPage statusCode={404} title="Action not found" />
-      </SizingWrapper>
-    );
-  }
-
-  return (
-    <SizingWrapper gridClassName="order-1 pt-6 pb-6 md:pb-10">
-      {loading ? (
-        <div>Loading...</div>
-      ) : (
-        <ActionDangerZone
-          actionIdentifier={action!.action}
-          onDelete={handleDelete}
-          isDeleting={isDeleting}
-          canDelete={true}
-        />
-      )}
-    </SizingWrapper>
-  );
-};
