--- web/scenes/PortalV3/Teams/TeamId/Apps/AppId/WorldIdActions/ActionId/Settings/page/index.tsx
@@ -1,73 +1,0 @@
-"use client";
-import { use } from "react";
-import { ErrorPage } from "@/components/ErrorPage";
-import { SizingWrapper } from "@/components/SizingWrapper";
-import { TryAction } from "@/scenes/PortalV3/Teams/TeamId/Apps/AppId/Actions/ActionId/Settings/TryAction";
-import Skeleton from "react-loading-skeleton";
-import { useGetSingleActionV4Query } from "@/scenes/common/Teams/TeamId/Apps/AppId/WorldIdActions/ActionId/page/graphql/client/get-single-action-v4.generated";
-import { UpdateActionV4Form } from "../UpdateActionV4Form";
-import { adaptActionV4ForTryAction } from "./utils/adapt-action-v4";
-
-type WorldIdActionIdSettingsPageProps = {
-  params: Promise<Record<string, string>>;
-};
-
-export const WorldIdActionIdSettingsPage = (
-  props: WorldIdActionIdSettingsPageProps,
-) => {
-  const params = use(props.params);
-  const actionId = params?.actionId;
-  const teamId = params?.teamId;
-  const appId = params?.appId;
-
-  const { data, loading, error } = useGetSingleActionV4Query({
-    variables: { action_id: actionId ?? "" },
-  });
-
-  const action = data?.action_v4_by_pk;
-
-  if (error) {
-    return (
-      <SizingWrapper gridClassName="order-1 md:order-2">
-        <ErrorPage statusCode={500} title="Failed to load action" />
-      </SizingWrapper>
-    );
-  }
-
-  if (!loading && !action) {
-    return (
-      <SizingWrapper gridClassName="order-1 md:order-2">
-        <ErrorPage statusCode={404} title="Action not found" />
-      </SizingWrapper>
-    );
-  }
-
-  return (
-    <SizingWrapper gridClassName="order-1 pt-6 pb-6 md:pb-10">
-      <div className="grid w-full grid-cols-1 items-start justify-between gap-x-32 gap-y-10 md:grid-cols-1fr/auto">
-        {loading ? (
-          <Skeleton count={4} />
-        ) : (
-          <UpdateActionV4Form action={action!} appId={appId ?? ""} />
-        )}
-
-        {loading ? (
-          <Skeleton className="md:w-[480px]" height={400} />
-        ) : (
-          <TryAction
-            action={adaptActionV4ForTryAction({
-              action: action!.action,
-              description: action!.description,
-              environment: "production",
-              rp_registration: {
-                app_id: action!.rp_registration.app_id,
-              },
-            })}
-            is_v4_action={true}
-            enableKiosk={false}
-          />
-        )}
-      </div>
-    </SizingWrapper>
-  );
-};
